requirements.txt
pandas>=2.0.0
playwright>=1.40.0
yfinance>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
pyarrow>=10.0.0
numpy<2.4
numba>=0.63.1
llvmlite>=0.46
notebook>=7.0.0
ipykernel>=6.30.0
tqdm>=4.66.0
matplotlib>=3.8.0
scipy>=1.13.0
ipywidgets>=8.1.0
jupyterlab-widgets>=3.0.9